CafeProjectJava
An automation project that will make the cafe manager's job easier. The administrator can add the products in his cafe to the automation. When the order comes, the cafe manager can select the registered products and their numbers according to the order, see the total price and give the customer a receipt.